requirements.txt
mysql-connector-python==8.0.29
Kivy==2.1.0
argon2-cffi==21.3.0
